requirements.txt
# Core Data Science & ML Libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Database
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0

# OpenAI & AI Libraries (Level 5)
openai>=1.0.0
python-dotenv>=1.0.0

# Optional: Advanced AI tools
# langchain>=0.1.0
# chromadb>=0.4.0

# API Dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
slowapi>=0.1.0
python-multipart>=0.0.6

